rename logger -> _logger
